+++ packages/react-native-the-sheet/src/bottom-sheet/types.ts
@@ -37,6 +37,7 @@ export type BottomSheetProps = PropsWithChildren & {
   snapPoints?: SnapPoint[]
   enableFloat?: boolean
   enableOverdrag?: boolean
+  disableDrag?: boolean
 
   fill?: boolean