+++ packages/react-native-embedded-stack-navigator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # react-native-embedded-stack-navigator
 
+## 2.0.3
+
+### Patch Changes
+
+- Add disableDrag to BottomSheet
+
 ## 2.0.2
 
 ### Patch Changes